backend/requirements.txt
fastapi==0.115.11
motor==3.7.0
passlib==1.7.4
pydantic==2.10.6
pymongo==4.11.2
python-dotenv==1.0.1
python_jose==3.4.0
